Move FailedCfd out of sql
The Sql module should not export any types. This will allow us in the future to have a different sql module such as postgres.
